Got side bar working correctly
