+++ Micro-Grid Management/ClientApp/public/favicon.ico
