+++ Micro-Grid Management/ClientApp/src/App.jsx
@@ -1,8 +1,6 @@
 import React, {Component} from 'react';
 import './css/custom.css';
 import {Routes, Route} from "react-router-dom";
-import {Home} from "./pages/Home"
-import {About} from "./pages/About"
 import {NavBar} from "./components/NavBar";
 import {Simulate} from "./pages/Simulate";
 import {History} from "./pages/History";
@@ -12,17 +10,15 @@ export default class App extends Component {
 
     render() {
         return (
-            <>
+            <div className={"app"}>
                 <NavBar />
-                <div>
+                <div className={"app"}>
                     <Routes>
-                        <Route path={"/"} element={<Home/>}></Route>
-                        <Route path={"/about"} element={<About/>}></Route>
-                        <Route path={"/simulate"} element={<Simulate/>}></Route>
+                        <Route path={"/"} element={<Simulate/>}></Route>
                         <Route path={"/history"} element={<History/>}></Route>
                     </Routes>
                 </div>
-            </>
+            </div>
         );
     }
 }
