+++ Micro-Grid Management/ClientApp/src/css/custom.css
@@ -1,25 +1,38 @@
 
 
-.btn-primary {
-  color: #fff;
-  background-color: #1b6ec2;
-  border-color: #1861ac;
-}
-
-#nav {
-  padding: 233px;
-}
-
-.form-label{
-  padding-right: 12px;
-  font-size: 15px;
-}
-
-#Home{
-  width: 100vw;
-  height: 100vh;
-}
-
 .side-bar{
-  align-content: flex-start;
+  align-content: center;
+  width: 160px;
+  height: 100vh;
+  background: #3F4856;
 }
+
+.navbar{
+  background: #585664;
+}
+
+li{
+  color: white;
+}
+
+.app{
+  background: #D9D9D9;
+}
+
+form{
+  background: #585664;
+  width: 490px;
+  height: 300px;
+  padding-left: 40px;
+  color: white;
+  align-content: center;
+  display:grid;
+}
+
+.h5{
+  background: #3F4856;
+  height: 30px;
+  width: 30px;
+}
+
+
