--- Micro-Grid Management/ClientApp/src/pages/About.jsx
@@ -1,9 +1,0 @@
-﻿import {TabPanel} from "../components/SideBar";
-
-export function About(){
-    return(
-        <>
-            
-        </>
-    );
-}
